Tested Admin, Customer, Sale, and Log services

+++ customer_service/customer.py
@@ -5,7 +5,7 @@ import requests
 
 from customer_service.models import Customer, customer_schema, customers_schema
 from shared.db import db, ma, bcrypt
-from shared.token import create_token, ADMIN_PATH, extract_auth_token, decode_token, jwt, LOG_PATH
+from shared.token import create_token, ADMIN_PATH, extract_auth_token, decode_token, jwt, LOG_PATH, CUSTOMER_PATH
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = "sqlite:///lab-project.db"
@@ -38,7 +38,7 @@ def get_all_customers():
         abort(403, "Something went wrong")
 
     try:
-        admin = requests.get(f"{ADMIN_PATH}/admin/{user_id}")
+        admin = requests.get(f"{ADMIN_PATH}/admin:{user_id}")
 
         if admin.status_code == 404:
             return abort(401, "Unauthorized")
@@ -49,10 +49,11 @@ def get_all_customers():
         customers = Customer.query.all()
         return jsonify(customers_schema.dump(customers)), 200
     except Exception as e:
+        print(e)
         abort(500, "Server Error")
 
 
-@app.route('/customer/<string:full_name>', methods=['GET'])
+@app.route('/customer:<string:full_name>', methods=['GET'])
 def get_customer_by_name(full_name):
     '''
     Get customer by username.
@@ -108,8 +109,8 @@ def create_customer():
     password = request.json['password']
     age = request.json['age']
     address = request.json['address']
-    gender = request.json['gender']
-    marital_status = request.json['marital_status']
+    gender = request.json['gender'].upper()
+    marital_status = request.json['marital_status'].upper()
 
     if type(full_name) != str or type(username) != str or type(password) != str or type(age) != int or type(address) != str or type(gender) != str or type(marital_status) != str:
         abort(400, "Bad Request")
@@ -129,11 +130,12 @@ def create_customer():
 
         return jsonify(customer_schema.dump(customer)), 200
     except Exception as e:
+        print(e)
         abort(500, "Server Error")
 
 
-@app.route('/customer:<int:customer_id>', methods=['PUT'])
-def update_customer(customer_id):
+@app.route('/customer', methods=['PUT'])
+def update_customer():
     '''
     Update customer information.
 
@@ -154,6 +156,27 @@ def update_customer(customer_id):
         404: Not Found
         500: Server Error
     '''
+    
+    
+    token = extract_auth_token(request)
+    if not token:
+        abort(403, "Something went wrong")
+    try:
+        customer_id = decode_token(token)
+    except (jwt.ExpiredSignatureError, jwt.InvalidTokenError):
+        abort(403, "Something went wrong")
+
+    try:
+        admin = requests.get(f"{CUSTOMER_PATH}/customer:{customer_id}")
+
+        if admin.status_code == 404:
+            return abort(401, "Unauthorized")
+        
+        if admin.status_code == 500:
+            return abort(500, "Server Error")
+    except:
+        abort(500, "Server Error")
+    
     balance = request.json.get('balance')
     username = request.json.get('username')
     password = request.json.get('password')
@@ -176,7 +199,7 @@ def update_customer(customer_id):
         abort(400, "Bad Request")
 
     try:
-        customer = Customer.query.filter_by(customer_id=customer_id).first()
+        customer = Customer.query.filter_by(user_id=customer_id).first()
 
         if balance:
             customer.balance += balance
@@ -184,7 +207,7 @@ def update_customer(customer_id):
         if username:
             dup = customer.query.filter_by(username=username).first()
 
-            if dup and dup.customer_id != customer_id:
+            if dup and dup.user_id != customer_id:
                 return abort(400, "Username already taken")
             
             customer.username = username
@@ -193,7 +216,7 @@ def update_customer(customer_id):
             customer.address = address
 
         if marital_status:
-            customer.marital_status = marital_status
+            customer.marital_status = marital_status.upper()
 
         db.session.commit()
         
@@ -201,11 +224,12 @@ def update_customer(customer_id):
 
         return jsonify(customer_schema.dump(customer)), 200
     except Exception as e:
+        print(e)
         abort(500, "Internal Server Error")
 
 
-@app.route('/customer/<int:customer_id>', methods=['DELETE'])
-def delete_customer(customer_id):
+@app.route('/customer', methods=['DELETE'])
+def delete_customer():
     '''
     Delete Customer.
 
@@ -218,24 +242,62 @@ def delete_customer(customer_id):
         500: Server Error
     '''
     
+    token = extract_auth_token(request)
+    if not token:
+        abort(403, "Something went wrong")
     try:
-        customer = Customer.query.filter_by(customer_id=customer_id).first()
-
-        if not customer:
-            abort(404, "Customer Not Found")
-
-        db.session.delete(customer)
-        db.session.commit()
+        id = decode_token(token)
+    except (jwt.ExpiredSignatureError, jwt.InvalidTokenError):
+        abort(403, "Something went wrong")
         
-        requests.post(f"{LOG_PATH}/add-log", json={"message": f"Deleted customer: {customer.username}"})
+    response1 = requests.get(f"{CUSTOMER_PATH}/customer:{id}")
+    response2 = requests.get(f"{ADMIN_PATH}/admin:{id}")
+    
+    if response1.status_code == 200:
+        customer_id = response1.json()['user_id']
+        try:
+            customer = Customer.query.filter_by(user_id=customer_id).first()
 
-        return {"Message": "Customer Deleted"}
-    except Exception as e:
-        abort(500, "Server Error")
+            if not customer:
+                abort(404, "Customer Not Found")
+
+            db.session.delete(customer)
+            db.session.commit()
+            
+            requests.post(f"{LOG_PATH}/add-log", json={"message": f"Deleted customer: {customer.username}"})
+
+            return {"Message": "Customer Deleted"}
+        except Exception as e:
+            abort(500, "Server Error")
+            
+    elif response2.status_code == 200:
+        if "customer_id" not in request.json():
+            abort(400, "Bad Request")
+        customer_id = request.json()['customer_id']
+        
+        try:
+            admin_id = response2.json()['admin_id']
+            customer = Customer.query.filter_by(user_id=customer_id).first()
+
+            if not customer:
+                abort(404, "Customer Not Found")
+
+            db.session.delete(customer)
+            db.session.commit()
+            
+            requests.post(f"{LOG_PATH}/add-log", json={"message": f"Admin {admin_id} Deleted customer: {customer.username}"})
+
+            return {"Message": "Customer Deleted"}
+        except Exception as e:
+            abort(500, "Server Error")
+    
+    else:
+        abort(403, "Unauthorized")
+        
 
 
-@app.route('/deduct/<int:customer_id>', methods=['POST'])
-def deduct(customer_id):
+@app.route('/deduct', methods=['POST'])
+def deduct():
     '''
     Deduct from customer balance.
 
@@ -251,6 +313,26 @@ def deduct(customer_id):
         404: Customer Not Found
         500: Server Error
     '''
+    
+    token = extract_auth_token(request)
+    if not token:
+        abort(403, "Something went wrong")
+    try:
+        customer_id = decode_token(token)
+    except (jwt.ExpiredSignatureError, jwt.InvalidTokenError):
+        abort(403, "Something went wrong")
+
+    try:
+        admin = requests.get(f"{CUSTOMER_PATH}/customer:{customer_id}")
+
+        if admin.status_code == 404:
+            return abort(401, "Unauthorized")
+        
+        if admin.status_code == 500:
+            return abort(500, "Server Error")
+    except:
+        abort(500, "Server Error")
+    
     if 'amount' not in request.json:
         abort(400, "Bad Request")
 
@@ -260,7 +342,7 @@ def deduct(customer_id):
         abort(400, "Bad Request")
 
     try:
-        customer = Customer.query.filter_by(customer_id=customer_id).first()
+        customer = Customer.query.filter_by(user_id=customer_id).first()
 
         if not customer:
             return abort(404, "Customer Not found")
@@ -279,8 +361,8 @@ def deduct(customer_id):
         return abort(500, "Server error")
     
 
-@app.route('/charge/<int:customer_id>', methods=['POST'])
-def charge(customer_id):
+@app.route('/charge', methods=['POST'])
+def charge():
     '''
     Charge customer with amount.
 
@@ -294,6 +376,26 @@ def charge(customer_id):
         404: Customer Not Found
         500: Server Error
     '''
+    
+    token = extract_auth_token(request)
+    if not token:
+        abort(403, "Something went wrong")
+    try:
+        customer_id = decode_token(token)
+    except (jwt.ExpiredSignatureError, jwt.InvalidTokenError):
+        abort(403, "Something went wrong")
+
+    try:
+        customer = requests.get(f"{CUSTOMER_PATH}/customer:{customer_id}")
+
+        if customer.status_code == 404:
+            return abort(401, "Unauthorized")
+        
+        if customer.status_code == 500:
+            return abort(500, "Server Error")
+    except:
+        abort(500, "Server Error")
+    
     if 'amount' not in request.json:
         abort(400, "Bad Request")
 
@@ -303,7 +405,7 @@ def charge(customer_id):
         abort(400, "Bad Request")
 
     try:
-        customer = Customer.query.filter_by(customer_id=customer_id).first()
+        customer = Customer.query.filter_by(user_id=customer_id).first()
 
         if not customer:
             return abort(404, "Customer Not found")
@@ -319,7 +421,7 @@ def charge(customer_id):
         return abort(500, "Server error")
 
 
-@app.route('/customer/<int:customer_id>', methods=['GET'])
+@app.route('/customer:<int:customer_id>', methods=['GET'])
 def get_customer_by_id(customer_id):
     '''
     Get customer by id.
@@ -333,7 +435,7 @@ def get_customer_by_id(customer_id):
         500: Server Error
     '''
     try:
-        customer = Customer.query.filter_by(customer_id=customer_id).first()
+        customer = Customer.query.filter_by(user_id=customer_id).first()
 
         if not customer:
             return abort(404, "Customer Not Found")
@@ -359,7 +461,7 @@ def authenticate():
         if not bcrypt.check_password_hash(customer.hashed_password, password):
             return abort(401, "Unauthorized")
         
-        d = {"token": create_token(customer.customer_id)}
+        d = {"token": create_token(customer.user_id)}
 
         return jsonify(d), 200
     except Exception as e:
